feat: use cobra to make it become a cli
